[red-knot] Move the `ClassBase` enum to its own submodule
